Extracted MockUserDAO to be easier to import
